Risolti ulteriori errori inerendi alla disposizione dei file
