+++ src/Business_logic/Transaction.java
@@ -1,5 +1,9 @@
+package Business_logic;
 // Transaction.java
 // Abstract superclass Transaction represents an ATM transaction
+
+import Database.BankDatabase;
+import GUI.Screen;
 
 public abstract class Transaction
 {
