+++ src/Database/BalanceInquiry.java
@@ -1,5 +1,9 @@
+package Database;
 // BalanceInquiry.java
 // Represents a balance inquiry ATM transaction
+
+import Business_logic.Transaction;
+import GUI.Screen;
 
 public class BalanceInquiry extends Transaction
 {
