+++ src/Database/BankDatabase.java
@@ -1,5 +1,8 @@
+package Database;
 // BankDatabase.java
 // Represents the bank account information database 
+
+import GUI.Account;
 
 public class BankDatabase
 {
