+++ src/GUI/ATM.java
@@ -1,5 +1,11 @@
+package GUI;
 // ATM.java
 // Represents an automated teller machine
+
+import Business_logic.DepositSlot;
+import Business_logic.Transaction;
+import Database.BalanceInquiry;
+import Database.BankDatabase;
 
 public class ATM 
 {
