+++ src/GUI/Account.java
@@ -1,3 +1,5 @@
+package GUI;
+
 // Account.java
 // Represents a bank account
