+++ src/GUI/CashDispenser.java
@@ -1,3 +1,4 @@
+package GUI;
 // CashDispenser.java
 // Represents the cash dispenser of the ATM
