+++ src/GUI/Screen.java
@@ -1,23 +1,27 @@
-// Keypad.java
-// Represents the keypad of the ATM
-import java.util.Scanner; // program uses Scanner to obtain user input
+package GUI;
+// Screen.java
+// Represents the screen of the ATM
 
-public class Keypad
+public class Screen
 {
-   private Scanner input; // reads data from the command line
-                         
-   // no-argument constructor initializes the Scanner
-   public Keypad()
+   // displays a message without a carriage return
+   public void displayMessage( String message ) 
    {
-      input = new Scanner( System.in );    
-   } // end no-argument Keypad constructor
+      System.out.print( message ); 
+   } // end method displayMessage
 
-   // return an integer value entered by user 
-   public int getInput()
+   // display a message with a carriage return
+   public void displayMessageLine( String message ) 
    {
-      return input.nextInt(); // we assume that user enters an integer  
-   } // end method getInput
-} // end class Keypad  
+      System.out.println( message );   
+   } // end method displayMessageLine
+
+   // display a dollar amount
+   public void displayDollarAmount( double amount )
+   {
+      System.out.printf( "$%,.2f", amount );   
+   } // end method displayDollarAmount 
+} // end class Screen
