+++ src/GUI/Withdrawal.java
@@ -1,5 +1,9 @@
+package GUI;
 // Withdrawal.java
 // Represents a withdrawal ATM transaction
+
+import Business_logic.Transaction;
+import Database.BankDatabase;
 
 public class Withdrawal extends Transaction
 {
